refactor: Batch D — safe-partial command grouping (T-CMD-GROUP)
Reduce lib.rs without the risk of relocating ~200 Tauri command wrappers:

- Extract the pure (command-free) custom-font and dashboard-background media
  helpers into a new media.rs submodule (189 lines), re-exported with
  pub(crate) use so the thin command wrappers call them unchanged.
- Annotate the generate_handler! registry with 27 domain section headers
  (App lifecycle, Connections, Settings, SFTP, RDP, VNC, Dashboard, Installer,
  Watchdog, …). Comments are lexer-whitespace, so the command set is byte-for-
  byte identical — verified by a clean build.

lib.rs is now ~3,063 lines (from 3,222) and the command surface is navigable by
domain. Mass-moving the command wrappers into their domain modules was
deliberately deferred as higher-risk, lower-payoff.

Verified: cargo build + cargo test (614) and npm run check (lint + 187 tests + tsc) green.
