+++ src-tauri/src/media.rs
@@ -1,0 +1,192 @@
+#![allow(unused_imports)]
+use super::*;
+
+pub(crate) fn list_custom_fonts_sync() -> Result<Vec<CustomFontEntry>, String> {
+    let folder = custom_fonts_folder()?;
+    fs::create_dir_all(&folder).map_err(|error| {
+        format!(
+            "failed to create custom fonts folder {}: {error}",
+            folder.display()
+        )
+    })?;
+
+    let mut fonts = fs::read_dir(&folder)
+        .map_err(|error| {
+            format!(
+                "failed to read custom fonts folder {}: {error}",
+                folder.display()
+            )
+        })?
+        .filter_map(|entry| entry.ok())
+        .filter_map(|entry| custom_font_entry(entry.path()))
+        .collect::<Vec<_>>();
+
+    fonts.sort_by(|a, b| a.name.to_lowercase().cmp(&b.name.to_lowercase()));
+    Ok(fonts)
+}
+
+pub(crate) fn load_custom_font_data_sync(path: String) -> Result<CustomFontData, String> {
+    use base64::{Engine as _, engine::general_purpose::STANDARD};
+
+    let folder = custom_fonts_folder()?;
+    fs::create_dir_all(&folder).map_err(|error| {
+        format!(
+            "failed to create custom fonts folder {}: {error}",
+            folder.display()
+        )
+    })?;
+
+    let folder = folder
+        .canonicalize()
+        .map_err(|error| format!("failed to resolve custom fonts folder: {error}"))?;
+    let path = PathBuf::from(path);
+    let canonical_path = path
+        .canonicalize()
+        .map_err(|error| format!("failed to resolve custom font path: {error}"))?;
+
+    if !canonical_path.starts_with(&folder) {
+        return Err("custom font path must stay inside the fonts folder".to_string());
+    }
+
+    if custom_font_entry(canonical_path.clone()).is_none() {
+        return Err("custom font file must be .ttf, .otf, .woff, or .woff2".to_string());
+    }
+
+    let bytes = fs::read(&canonical_path).map_err(|error| {
+        format!(
+            "failed to read custom font {}: {error}",
+            canonical_path.display()
+        )
+    })?;
+
+    Ok(CustomFontData {
+        data_base64: STANDARD.encode(bytes),
+    })
+}
+
+pub(crate) fn custom_fonts_folder() -> Result<PathBuf, String> {
+    let exe_path = std::env::current_exe()
+        .map_err(|error| format!("failed to resolve app executable path: {error}"))?;
+    let exe_folder = exe_path
+        .parent()
+        .ok_or_else(|| "failed to resolve app executable folder".to_string())?;
+    Ok(exe_folder.join("fonts"))
+}
+
+pub(crate) fn backgrounds_folder() -> Result<PathBuf, String> {
+    let exe_path = std::env::current_exe()
+        .map_err(|error| format!("failed to resolve app executable path: {error}"))?;
+    let exe_folder = exe_path
+        .parent()
+        .ok_or_else(|| "failed to resolve app executable folder".to_string())?;
+    Ok(exe_folder.join("backgrounds"))
+}
+
+/// Best-effort: delete background media files no view references anymore.
+/// Never returns an error — cleanup failures must not break view mutations.
+pub(crate) fn prune_unreferenced_backgrounds(app: &tauri::AppHandle) {
+    let storage = app.state::<storage::Storage>();
+    let referenced = storage.with_connection_infallible(|conn| {
+        dashboard_storage::referenced_background_image_files(conn)
+            .map_err(|error| format!("{error:?}"))
+    });
+    let referenced = match referenced {
+        Ok(set) => set,
+        Err(error) => {
+            eprintln!("background prune skipped: {error}");
+            return;
+        }
+    };
+    let folder = match backgrounds_folder() {
+        Ok(folder) => folder,
+        Err(error) => {
+            eprintln!("background prune skipped: {error}");
+            return;
+        }
+    };
+    let entries = match fs::read_dir(&folder) {
+        Ok(entries) => entries,
+        Err(_) => return, // folder may not exist yet — nothing to prune.
+    };
+    for entry in entries.filter_map(Result::ok) {
+        let path = entry.path();
+        let name = match path.file_name().and_then(|n| n.to_str()) {
+            Some(name) => name.to_string(),
+            None => continue,
+        };
+        if !referenced.contains(&name) {
+            if let Err(error) = fs::remove_file(&path) {
+                eprintln!(
+                    "failed to prune background image {}: {error}",
+                    path.display()
+                );
+            }
+        }
+    }
+}
+
+pub(crate) fn custom_font_entry(path: PathBuf) -> Option<CustomFontEntry> {
+    if !path.is_file() {
+        return None;
+    }
+
+    let extension = path
+        .extension()
+        .and_then(|extension| extension.to_str())
+        .map(|extension| extension.to_lowercase())?;
+
+    if !is_supported_font_extension(&extension) {
+        return None;
+    }
+
+    let name = path
+        .file_stem()
+        .and_then(|name| name.to_str())
+        .or_else(|| path.file_name().and_then(|name| name.to_str()))?
+        .to_string();
+
+    Some(CustomFontEntry {
+        name,
+        path: path.to_string_lossy().into_owned(),
+        extension,
+    })
+}
+
+pub(crate) fn is_supported_font_extension(extension: &str) -> bool {
+    matches!(extension, "ttf" | "otf" | "woff" | "woff2")
+}
+
+/// Returns the lowercased extension if `path` is a supported background media file.
+pub(crate) fn background_media_extension(path: &std::path::Path) -> Option<String> {
+    let extension = path
+        .extension()
+        .and_then(|extension| extension.to_str())
+        .map(|extension| extension.to_lowercase())?;
+    if matches!(
+        extension.as_str(),
+        "png" | "jpg" | "jpeg" | "webp" | "gif" | "bmp" | "mp4" | "webm" | "mov" | "m4v" | "ogv",
+    ) {
+        Some(extension)
+    } else {
+        None
+    }
+}
+
+pub(crate) fn background_media_mime(extension: &str) -> &'static str {
+    match extension {
+        "png" => "image/png",
+        "jpg" | "jpeg" => "image/jpeg",
+        "webp" => "image/webp",
+        "gif" => "image/gif",
+        "bmp" => "image/bmp",
+        "mp4" | "m4v" => "video/mp4",
+        "webm" => "video/webm",
+        "mov" => "video/quicktime",
+        "ogv" => "video/ogg",
+        _ => "application/octet-stream",
+    }
+}
+
+pub(crate) fn background_media_extension_error() -> &'static str {
+    "background file must be .png, .jpg, .jpeg, .webp, .gif, .bmp, .mp4, .webm, .mov, .m4v, or .ogv"
+}
